Arreglado

+++ app/src/main/java/com/e/frutas_api/Fruit.kt
@@ -4,11 +4,11 @@ data class Fruit(
     val name:String,
     val id:Int,
     val family:String,
-    val carbohydrates:Int,
-    val protein: Int,
-    val fat:Float,
-    val calories:Int,
-    val sugar:Float,
+    val carbohydrates:Double,
+    val protein: Double,
+    val fat:Double,
+    val calories:Double,
+    val sugar:Double,
 
         )
